requirements_colab.txt
transformers
jinja2
pyyaml
diffusers
torch
piexif
accelerate
xformers
argcomplete
compel
safetensors
pytorch_lightning

numpy
opencv-python
onnx
onnxruntime
insightface
psutil
tk
customtkinter
pillow
tensorflow
opennsfw2
protobuf
tqdm
gfpgan